add App info > Storage > Manage storage button for apps that support it
There's a "Clear storage" button in App info > Storage & cache. When app implements a "manage
storage space" activity, that button opens it instead of actually clearing the storage.

"Manage space" activity might not provide an option to clear app storage or it might not work at
all.

This change adds a separate button to launch that activity and makes the "Clear storage" button
ignore its presence.

Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>